+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 more about how to use the Metrics API.
 - Pre-main profiling data is now attached to the app start transaction (#3736)
 - Release framework without UIKit/AppKit (#3793)
+- Add the option swizzleClassNameExcludes (#3813)
 
 ### Fixes